requirements.txt
python-telegram-bot==21.6
anthropic==0.34.2
groq
python-dotenv
